env: extend cpumask to hold more than 64 cpus
This patch replaces cpumask from uint64_t to the the spdk_cpuset_t
(typedef of cpu_set_t for Linux and cpuset_t for FreeBSD used for
setting affinity of threads to the cpus).
This structure allows to hold set of more than 64 cores (e.g. for now
it's 1024 cpus for Linux).

This patch also extends parsing function allowing to set cpumask
using a list of cpus e.g. "[0-4,10,12]" sets mask of 0,1,2,3,4,10,12.

Change-Id: I475c3ba7fab629021a22e03176e57e400dd24a49
Signed-off-by: Tomasz Kulasek <tomaszx.kulasek@intel.com>