Update patch set 1
Patch Set 1:

For the I/O freezing, we will need a new for_each_channel operation that allows each channel to complete its function asynchronously.  This will allow each channel to wait until all existing I/O are completed before proceeding to the next channel.

Ben needs this for NVMe-oF too though, so he's already working on implementing it.

Would love help in getting this implemented.  It is going to be a lot of work!

Patch-set: 1