scsi: remove lun name
There is no need to keep a lun name anymore - we always
use the bdev name as the lun name so it is not providing
any additional value.  This also keeps us from associating
the same bdev with different LUNs on different iSCSI target
nodes or vhost-scsi controllers.

Side effects of this change are:

1) Remove the get_luns() RPC.  Not very useful and more
   functionality exists in get_bdevs() anyways.
2) Remove the entire lun_db.c file, since there is no
   longer a need to track or lookup based on a LUN
   name anymore.
3) Use "bdev_name" across the APIs to make it more clear
   what these names refer to.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I3d42fde22087352ce1d5dc80178bd8c5cac8cb7c