Update patch set 6
Patch Set 6:

I think we have the same problem with the poll groups too if you ctrl-c half way through initializing them. I think the only way to deal with this is to make the regular shutdown path check if the acceptor poller is running before shutting it down and check if each poll group exists before destroying it. I think for the poll groups it just asserts today.

Patch-set: 6