Update patch set 9
Patch Set 9: Code-Review-1

I have a problem with DPDK not here.

If the DPDK NIC PMD (virtio-user-net) is used along wit SPDK NVMe (virtio-user-(scsi|blk)) this patch will make:
1. SPDK vhost fatality - memory may be already registered by DPDK virtio-net driver and we will be not abble to register it again (vtophys() will fail)
2. The DPDK virtio-net fatality - if we are lucky enought to start with some (say just one) NVMe device and then hotremove it we will unregister IOMMU memory that DPDK virtio-user-net driver still need.

So I think the only way to fix this one and for good is to upstream all this VFIO stuff from here to DPDK.
Let talk about this on monday.

Patch-set: 9