Update patch set 7
Patch Set 7:

The distribution of connections across cores is now done in the application on a per-connection basis (inside the new_qpair callback). We'll need to give the user the ability to at least ask what NUMA node the NIC is on when they do this. I haven't added that API just yet.

Unfortunately, there is no way to know what SSD a connection is for until too late, so the user won't be able to take NUMA into consideration on the storage side for RDMA. For other transports, they have more flexibility.

Patch-set: 7
Label: Code-Review=0