Update patch set 5
Patch Set 5:

Thank for the explanation.  Would be optimal if we could detect # of disks and skip the tests in question if there is only one disk.  Next best, better than commenting I think, is to just define define a parm at the top of the file for number of available disks in test system or something and then use that to decide whether to run the multi disk test or not and if not print a big message that says something like "skipping test XYZ because only one disk is present" as sort of a better reminder than a comment that we have a hardocded limitation that we should alter later

Patch-set: 5
Label: Code-Review=0