Update patch set 1
Patch Set 1: Code-Review+2

This is a step in the right direction, but if request_transfer_out returns -1 the code currently just asserts. We should be correctly handling this by not doing the data transfer and attempting to send a completion.

If we can't send a completion, I don't know what the code should do. Retry maybe, but that may not work either.

Patch-set: 1
Label: Code-Review=+2