Update patch set 7
Patch Set 7: Code-Review-1

(1 comment)

Overall this looks OK - but a 366 line patch to a single file is pretty big.  :)  Surely there is some way you could break this up a bit?

My main concern is how do we find out how much data was actually transferred for a completed command?  This used to be in the data_transferred field in virtio_req.  I know we're not using that today, but as SPDK gets more sophisticated we'll want to handle underflow conditions.  Also if we ever want to report full INQUIRY data (and not just first 36 bytes of basic INQUIRY data) we'll need to know how much the target actually filled out.

Patch-set: 7
Label: Code-Review=-1