Update patch set 19
Patch Set 19: Code-Review-1

(2 comments)

I'm not sure we should check this patch in without the actual thin provision code.  Otherwise if this patch goes in, it will seem you can create thin provisioned blobs.

Or if we want to put this in now, please add to the comment, saying something like "this only adds the option and metadata flags - actual functionality will be added in an upcoming commit".

Please also get the xattr spdk_blob_opts patch before this patch.  I made that request in the xattrs patch but mentioning it here too now.

Patch-set: 19
Label: Code-Review=-1