Update patch set 5
Patch Set 5: Code-Review-1

This is not against Cunyin's patch. We still have issues for the unexpected exit for bdeveperf. If we set a large time, and we exit early. We still need to solve it before change nvme perf to this mode, otherwise cannot test a long time also for nvme perf.

Patch-set: 5
Label: Code-Review=-1