Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> Does this document belong in project spdk/spdk or spdk/spdk.github.io? I was trying to view the doc on our website and I could not find the link in any of the builds. Usually when I have made changes to docs they go to spdk.github.io so that we can review how they will render on the website.

As far as I can see, this document belongs in spdk/spdk. spdk/spdk.github.io contains a precompiled version of docs and has to be updated separately. I've tried to compile the docs for preview in patch 391906 [1], but due to different environment/doxygen version it looks a bit... off. [2]

[1] https://review.gerrithub.io/c/391906/
[2] https://ci.spdk.io/spdk.github.io/builds/review/746d2463110e7db166c584811b0e8c199a212076.1513328878/spdk.intel.com/_site/doc/vhost.html

Patch-set: 2
Label: Code-Review=0