[RFC] vpp: add support for vpp comm library
This patch replaces current sock.c with vpp.c when
configured with flag "--with-vpp".

In such case iscsi target attaches to VPP instance and
data is transfered using VPP instead of kernel TCP/IP stack.

Currently this is just a proof of concept, that attaching
and communicating through VPP should be possible in this manner.
Next step is to change to VPP session API that is more event driven.
This would include adding interface between iscsi_tgt and network
stack that is universal for both Socket API and VPP session API.

Change-Id: I4c2945c76878dbf24ff2f8612e0a21615f7d87e6
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>