Update patch set 6
Patch Set 6:

This is really needed to guard against one thread deleting a blob and then another thread (or even the same thread I guess) opening that same blob while the deletion is in progress.

Currently, without this patch - the only way to find out if a blob is valid is to read the metadata page associated with the specified blob ID.  So if a deletion is in progress, the blob could be opened and then the metadata page could be cleared destroying the blob.

With this array, now we can clear the bit in the array when the deletion process starts.  When a thread tries to open a blob, it looks in the blobid array to see if its a valid blobid - if it isn't, it won't even try to read the metadata page.

Patch-set: 6