Update patch set 8
Patch Set 8: Code-Review-1

(12 comments)

First, thanks for tackling this - it is definitely a good cleanup and helps support larger systems.

As a general note, I'm not totally sold on tying the implementation of the SPDK cpuset type to the system's cpu_set_t; this limits us to the fixed-size array that the system headers choose, and if we later want to increase this (e.g. to use Linux's CPU_ALLOC, CPU_SET_S, etc. functions), we'd have to break the API again, since the size of our cpuset is part of the API/ABI.  It might be a good idea to make our cpuset a pointer to a dynamically allocated struct with its contents hidden in the cpuset.c implementation.  However, I'll defer to the other core maintainers on this matter if there are differing opinions.

Patch-set: 8
Label: Code-Review=-1