test/lvol: wait for lvolstore tasting after nvme bdev construction
A couple of the lvol tests delete the nvme bdev, then construct
it again to test lvolstore tasting.  But constructing the nvme
bdev and then the tasting can take a tiny bit of time, and if
we check for the lvolstore too soon, it could fail.  There are
a large number of intermittent test failures in the test pool
that show this signature.

So sleep one second to allow time for the tasting to complete.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: Iebabe09458302d8300e704f232fdbb245b562268