[RFC] blob: add locking
This would enable metadata operations from any thread
simultaneously.

I really hate where this patch is going though - the
locking gets messy and I haven't covered everything
yet.  Getting this out for review so others can look
at it.

I'm really thinking instead of keeping the API such
that metadata operations can only be performed on the
thread that inits or opens the blobstore.  We can
use the spdk_thread API to enforce it - something we
did not have when blobstore first went into the tree.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I3a3965295a6e1712b76afffd89cc555996d6a393