Update patch set 6
Patch Set 6:

Thank you for your effort and this will be helpful.

I have a simple question, why the following is the first?
  1 Do not pay efforts on detailed iSCSI protocol parsing handling.

I may not have enough expertise in iSCSI protocol but it looks that current SPDK iSCSI target is poor and do not have clean implementation.

This may be due to istgt on which SPDK iSCSI target is based.

For example
- I cannot see enough protocol compliance in istgt's state transition management.
- data overflow/underflow management is wrong.

I'm not hurry in doing the above but "Do not pay efforts ..." is not too extreme? These 3 ideas are applied to NVMf target too?

Patch-set: 6