Update patch set 6
Patch Set 6:

Paul have created the similar docs.
https://review.gerrithub.io/#/c/384118/

But that is higher level than this. this is very detailed. Creating this just once is good but I'm afraid that we will not be able to maintain this. As many have admitted already, iSCSI's APIs are old but not clean yet. iSCSI with VPP is on-going and remains to be done. I'm still adding function and refactoring code.

I'm positive for documentation effort and I would like to hear Jim's thought.

Patch-set: 6