Update patch set 6
Patch Set 6:

I replied some comments but I'm reading and this is very well written. Thank you so much.

Your docs is about northbound API as you write that this is for application programmer utilizing SPDK iSCSI.

It will be better if you can add anything to show this purpose in the doc name.

Why I replied this is that our production team are interested in southbound API too and they will expect it is contained.

Patch-set: 6