Update patch set 6
Patch Set 6:

I understand that we try to develop SPDK such that each library can be used independently. This doc describes northbound API and it looks that this assumes SPDK SCSI, BDEV, NET are also used.

It may be helpful for any reader if there is any description about these.

I'm gradually understanding this doc and sorry for any misconception.

Patch-set: 6