test/nvmf: shorten nightly nvmf+pmem test
This test ran for 10 minutes, which is half of the allotted time for the
entire nightly test run.

While we're here, remove some of the duplication - we don't need to run
both the nightly long run and the shorter normal test run of the same
test, and they should be grouped together.

Change-Id: I66385054d425fb8b032e282050fd39e927a8eb80
Signed-off-by: Daniel Verkamp <daniel.verkamp@intel.com>