lvol, blob: make channel opts configurable for blob.
We need to make the channel configurable for blob.
Reason, for iSCSI tests, if there is one CPU core,
there will be only one channel, thus read tests would
fail since we need more opts.

Current solution is a temp solution, I think the very
exact fix is that we need to restrict sending I/Os
to the blob, if there is no channel opts.

Change-Id: I211f7a89d144af2c96ad4cc1bd7ac8e94adc72e7
Signed-off-by: Ziye Yang <optimistyzy@gmail.com>