Update patch set 27
Patch Set 27: Code-Review+1

Well, the test file doesn't show up where I expected it and its different for my test system where its in $testdir.  As long as its not a problem that it is created in the current working dir, deleting it from there works fine.

Patch-set: 27
Label: Code-Review=+1