Update patch set 16
Abandoned

Abandon this. We have two proposed method to do this:

1. For those already configured bdev in the configuration file. Added the [QoS] section to assign the QoS at the creation time.

2. Enable the QoS when constructing the Bdev via RPC no matter what kind of backend device it is. This can be completed in two steps 1) create the bdev 2) enable the qos on this bdev

If we do need to create the bdev and enable the QoS at the same time for the single RPC method. We can add it later.

In this way, there is no need to spread the QoS setting to each kind of Bdev types via RPC methods.

Patch-set: 16
Status: abandoned
Tag: autogenerated:gerrit:abandon