bdev/qos: update the RPC method to display the actual IOPS
When QoS is enabled on the bdev, only the master thread will operate
the IOs. Add the actual QoS calculation in that periodically running
poller dynamically instead of having the average IOPS for the overall
running time.

Update the RPC method to display this timely actual IOPS when QoS has
been enabled on the bdev.

Change-Id: I80ff89b6cd424691e087ecfd1c6605625d03fff7
Signed-off-by: GangCao <gang.cao@intel.com>