Update patch set 10
Patch Set 10: Code-Review-1

I'm having a hard time figuring out how we are testing multisocket with this patch.  To me this all just looks like refactoring.

Probably either the commit message needs to be updated, or the refactoring should be broken out into a separate patch to make it more clear what the multisocket changes are.

Patch-set: 10
Label: Code-Review=-1