Update patch set 6
Patch Set 6:

Hi Shuhei,

We have the plan to provide programming docs (including all components) to users. In my mind, we should provide more details to customers for programming docs. And currently, all my talked for iSCSI is the public API exported in lib/iscsi/*.h, even if we changed the API, it is still not a huge efforts. If we change the API Function, we should change the documents to maintain the consistency.

Patch-set: 6
Label: Code-Review=0