iscsi: restore data_transferred accumulation for read
In the patch 393582, by analyzing the code,
accumulation of write completion to bdev was duplicated by
data_transferred and bytes_completed.

Hence accumulation of data_transferred for write was intended
to be removed.

However by mistake accumulation of data_transferred for read
was removed.

In this patch restore accumulation of data_transferred for
read and remove accumlation of data_transferred for write.

Test code to catch this degradation will be added by another
patch soon.

Change-Id: Iea9883e8ef1bfb0bdf00e291712e6faf2fad281f
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Reviewed-on: https://review.gerrithub.io/393713
Tested-by: SPDK Automated Test System <sys_sgsw@intel.com>
Reviewed-by: Daniel Verkamp <daniel.verkamp@intel.com>
Reviewed-by: Ziye Yang <optimistyzy@gmail.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>