iscsi: data_out_cnt of task and conn is not used anymore
Large read is splited into multiple payloads and max size of read
payload is 64KB and data_in_cnt is used to count the number of
read payloads. data_in_cnt is necessary to limit the number of
read payload per connection.

Large write is also splited into multiple payloads. It looks that
data_out_cnt is intended to be used to count the number of 64KB
write payload but current max size of write payload is 64KB * 16
and it looks that data_out_cnt does not serve any function anymore.

This is found through developing UT code to validate primary-subtask
relationship in iSCSI.

Change-Id: I17eeac74626d73ea5ffb2c09fb116dcff0bc2650
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>