Update patch set 26
Patch Set 26: Code-Review-1

Let's keep patch separate for now.  We can look at maybe merging them after review is done.

I'd like to understand more about the dirty shutdown sequence.  How do we write out cluster index 0 for normal (not thin provisioned) blobs?  And if a blob is not thin provisioned, and we get a cluster index of 0, we have to return a failure - it is not valid.

Marking -1 until we bottom out on the dirty shutdown case you mentioned.

Patch-set: 26
Label: Code-Review=-1