Update patch set 5
Patch Set 5:

Hi Shuhei,
It seems that currently data_out_cnt of conn is not well used. For the data_out_cnt, It seems that it records the total number of data packet number related with a single out standing task.

Can we not remove it, but do the following things:

When a connections drops, check both data_out_cnt/data_in_cnt values, to track the resource free is correct?

Patch-set: 5