Update patch set 7
Patch Set 7:

I still need to review this but wanted to add some commentary first.

iSCSI hotplug is not really functional right now - it does not register any hotplug callback functions.  I tried adding them which fixes one problem but then others pop up - primarily that large writes waiting for R2Ts need to be cleaned up and there's no easy way to map the SCSI LUN that was removed back to all of the R2Ts for the connections that are associated with iSCSI target nodes containing that LUN.

I almost think that we need some kind of open() operation on SCSI devices/LUNs, like we do for bdevs.  Currently the SCSI hot remove callbacks are passed during construction which doesn't work (and that's not how the bdev layer does it).

So I'm not saying your patch here isn't good (I haven't reviewed it) but that we may need to rethink how LUN removal works entirely.

Patch-set: 7
Label: Code-Review=0