Update patch set 5
Patch Set 5:

The current proposal is to have the rate limiting on the single bdev. I've already added the different methods to set this QoS goal at the beginning if the application has a clear requirement via the configuration file or set/update the QoS goal the runtime through the RPC method.

I've also updated the bdevperf tool to test different situations and now the QoS control can take effect immediately with the current method.

This is the QoS (rate limiting) that this work is targeting for and to meet the real user customer requirements that we collected. For your concern of global / local, it may be another story to take care and as long as we can have a concise control on the single bdev. We can work with other cases like multiple bdevs, multiple cores situation easily.

For the nomem_io case, it is to cover the general situation where the system meets this case and QoS in this situation needs to properly co-work.

Patch-set: 5