Update patch set 5
Patch Set 5:

Thank you Gang, it may be better for us to accumulate discussion into Trello.

About your patch, the following is very complex and not extensible for me:

  2. Once the memory resource is enough (we have a threshold there), the IOs already queued on the nomem IO list will be drained one by one and they will be limited by the QoS policy (looks like this is not added in this patch yet)

This point is the foundational one and very important. What is the advantage of yours? I'm sorry but I cannot find yet. 

On the contrary, what is the disadvantage of mine?
- if we check qos first, only IOs allowed by qos were submitted and
- draining each nomem_io is already implemented and hence just utilize the existing mechanism for nomem_io.

I do not have any intention to repeat the same question but I have not got your explicit feedback about this yet. You have just updated the patch.

I believe this discussion will achieve the good fruit.

Patch-set: 5