Update patch set 4
Patch Set 4:

I spoke with Jim. In order to fix this in a performant way, we'd need to change the way I/O channels are created by the user (per blob instead of per thread). For now, go ahead and proceed with this patch. It will only break in reality for scenarios that we aren't targeting right now, like an implementation of bs_dev that goes straight to the nvme library.

Patch-set: 4