Update patch set 9
Patch Set 9: Code-Review-1

(3 comments)

Overall, I'm OK with this patch.  I think this might also fix the intermittent failures in the test pool on fedora-07 with the vhost initiator script.  I'm checking that now.

Regarding LUN open() - open() is not just about sharing a LUN, but also about explicitly telling the SCSI layer when it is done with the LUN (via a corresponding close()).  Trying to deduce it based on outstanding number of I/O doesn't work in some use cases - especially iSCSI where we need to do some amount of queuing to the LUN above the SCSI layer.

Patch-set: 9
Label: Code-Review=-1